move code into types.py
