--- effectful/__init__.py
@@ -1,1 +1,0 @@
-import effectful.internals.pretty  # noqa: F401
